Rename Package & add Access Modifier
