+++ README.md
@@ -2,3 +2,8 @@
 
 초기에 Zapier Integration을 위한 REST 기반 Web Application이었으나,
 해당 프로젝트 종료 후, Kotlin Toy Project로 사용하고자 함
+
+---
+##Project 구성
+- RDB - Oracle (Docker Image - oracle11g jaspeen/oracle-xe-11g) // Local
+ ---
