+++ src/main/kotlin/com/example/app/ZapappApplication.kt
@@ -1,11 +1,11 @@
-package com.example.zapapp
+package com.example.app
 
 import org.springframework.boot.Banner
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
 
 @SpringBootApplication
-class ZapappApplication
+open class ZapappApplication
 
 fun main(args: Array<String>) {
     runApplication<ZapappApplication>(*args) {
