+++ src/main/kotlin/com/example/app/repository/MemberRepository.kt
@@ -1,6 +1,6 @@
-package com.example.zapapp.Repository
+package com.example.app.repository
 
-import com.example.zapapp.Entity.Member
+import com.example.app.entity.Member
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
